Update cart controller - add remove from cart functionality - switch int with float for prices Update paypal controller - create order with new created entity after capture Update cart entity - switch int with float Update cart item entity - switch int with float Update paypal service - change capture method return type Update paypal interface Update cart template - add form for removing cart item - add form for checkout Update admin nav - add link for new create admin order controller Update templates Add migrations Create new paypalform trait Create new templates

+++ src/Entity/Cart.php
@@ -22,7 +22,7 @@ class Cart
     private $id;
 
     /**
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      */
     private $total = 0;
 
@@ -41,12 +41,12 @@ class Cart
         return $this->id;
     }
 
-    public function getTotal(): ?int
+    public function getTotal(): ?float
     {
         return $this->total;
     }
 
-    public function setTotal(?int $total): self
+    public function setTotal(?float $total): self
     {
         $this->total = $total;
